+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 **Other:**
 - Resolve Xcode 10.2 warnings with Swift 4.2.2 and 5.0 (#397) - @mjarvis
+- Update Swift Package Manager support (#403) - @Dschee
 
 # 4.1.1